+++ src/Response/Sender/Native.php
@@ -1,9 +1,11 @@
 <?php
 declare(strict_types = 1);
 
-namespace Innmind\Http;
+namespace Innmind\Http\Response\Sender;
 
 use Innmind\Http\{
+    Response,
+    Header\Custom,
     Header\Date,
     Header\SetCookie,
     Exception\LogicException,
@@ -18,7 +20,7 @@ use Innmind\Immutable\{
     SideEffect,
 };
 
-final class ResponseSender implements Response\Sender
+final class Native implements Response\Sender
 {
     public function __construct(
         private Clock $clock,
@@ -58,7 +60,7 @@ final class ResponseSender implements Response\Sender
                 return;
             }
 
-            if ($header instanceof Header\Custom) {
+            if ($header instanceof Custom) {
                 $header = $header->normalize();
             }