+++ CHANGELOG.md
@@ -80,6 +80,7 @@
 - `Innmind\Http\Header\Value` constructor is now private, use `::of` named constructor
 - `Innmind\Http\Header\Parameter` constructor is now private, use `::of` named constructor
 - `Innmind\Http\Sender` as been moved to `Innmind\Http\Response\Sender`
+- `Innmind\Http\ResponseSender` as been moved to `Innmind\Http\Response\Sender\Native`
 
 ### Removed